Oxorio code review 3

+++ contracts/interfaces/IAggregatorBase.sol
@@ -2,7 +2,7 @@
 pragma solidity ^0.8.2;
 
 interface IAggregatorBase {
-    
+
     /* ========== STRUCTS ========== */
 
     struct OracleInfo {
@@ -13,7 +13,7 @@ interface IAggregatorBase {
     }
 
     struct BlockConfirmationsInfo {
-        uint256 count; // count of submissions in block
+        uint8 count; // count of submissions in block
         bool requireExtraCheck; // exceed confirmation count for all submissions in block
         mapping(bytes32 => bool) isConfirmed; // submission => was confirmed
     }
@@ -22,7 +22,7 @@ interface IAggregatorBase {
 
     event AddOracle(address oracle, address admin, bool required); // add oracle by admin
     event UpdateOracle(address oracle, bool required, bool isValid); // update oracle by admin
-    event UpdateOracleAdmin(address oracle, address admin); // update oracle by oracle's admin 
+    event UpdateOracleAdmin(address oracle, address admin); // update oracle by oracle's admin
     event UpdateOracleAdminByOwner(address oracle, address admin); // update oracle by admin
     event DeployApproved(bytes32 deployId); // emitted once the submission is confirmed by min required aount of oracles
     event SubmissionApproved(bytes32 submissionId); // emitted once the submission is confirmed by min required aount of oracles
