+++ contracts/transfers/SignatureAggregator.sol
@@ -7,7 +7,7 @@ import "../periphery/WrappedAsset.sol";
 import "../libraries/SignatureUtil.sol";
 
 contract SignatureAggregator is AggregatorBase, ISignatureAggregator {
-    
+
     using SignatureUtil for bytes;
     using SignatureUtil for bytes32;
 
@@ -20,8 +20,8 @@ contract SignatureAggregator is AggregatorBase, ISignatureAggregator {
 
     /// @dev Constructor that initializes the most important configurations.
     /// @param _minConfirmations Common confirmations count.
-    function initialize(uint256 _minConfirmations) 
-        public initializer 
+    function initialize(uint8 _minConfirmations)
+        public initializer
     {
         AggregatorBase.initializeBase(_minConfirmations);
     }
@@ -95,13 +95,13 @@ contract SignatureAggregator is AggregatorBase, ISignatureAggregator {
     }
 
     /* ========== VIEW ========== */
-    
+
     /// @dev Returns whether transfer request is confirmed.
     /// @param _submissionId Submission identifier.
     /// @return _confirmations number of confirmation.
     /// @return _confirmed Whether transfer request is confirmed.
     function getSubmissionConfirmations(bytes32 _submissionId)
-        external view override returns (uint256 _confirmations, bool _confirmed)
+        external view override returns (uint8 _confirmations, bool _confirmed)
     {
         SubmissionInfo storage submissionInfo = getSubmissionInfo[
             _submissionId
